requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-qdrant>=0.2.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pypdf>=5.0.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
qdrant-client>=1.12.0
openai>=1.60.0
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
cryptography>=42.0.0
bcrypt>=4.0.0
python-multipart>=0.0.7
